4 sum

+++ Arrays/fourSum.cpp
@@ -18,7 +18,7 @@ public:
                 int k = j + 1;
                 int l = n - 1;
                 while (k < l) {
-                    int sum = nums[i] + nums[j] + nums[k] + nums[l];
+                    long long  sum = (long long)nums[i] + nums[j] + nums[k] + nums[l];
                     if (sum == target) {
                         vector<int> temp = {nums[i], nums[j], nums[k], nums[l]};
                         ans.push_back(temp);
